requirements.txt
Django>=2.1.3
djangorestframework>=3.9.0
flake8>=3.6.0
requests